Move more specimen code from study into specimen
