+++ specimen/src/org/labkey/specimen/SpecimenRequestException.java
@@ -1,4 +1,4 @@
-package org.labkey.api.specimen;
+package org.labkey.specimen;
 
 public class SpecimenRequestException extends RuntimeException
 {
