+++ specimen/src/org/labkey/specimen/actions/ManageRequestBean.java
@@ -3,7 +3,6 @@ package org.labkey.specimen.actions;
 import org.labkey.api.data.ActionButton;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.MenuButton;
-import org.labkey.api.specimen.SpecimenRequestStatus;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
@@ -13,6 +12,7 @@ import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.DisplayElement;
 import org.labkey.api.view.ViewContext;
 import org.labkey.specimen.SpecimenRequestManager;
+import org.labkey.specimen.SpecimenRequestStatus;
 import org.labkey.specimen.actions.SpecimenController.ImportVialIdsAction;
 import org.labkey.specimen.actions.SpecimenController.OverviewAction;
 import org.labkey.specimen.actions.SpecimenController.RemoveRequestSpecimensAction;
