+++ specimen/src/org/labkey/specimen/actions/SpecimenController.java
@@ -86,15 +86,11 @@ import org.labkey.api.security.permissions.ReadPermission;
 import org.labkey.api.security.permissions.UpdatePermission;
 import org.labkey.api.specimen.SpecimenMigrationService;
 import org.labkey.api.specimen.SpecimenQuerySchema;
-import org.labkey.api.specimen.SpecimenRequestException;
-import org.labkey.api.specimen.SpecimenRequestStatus;
 import org.labkey.api.specimen.SpecimenSchema;
 import org.labkey.api.specimen.Vial;
-import org.labkey.api.specimen.importer.RequestabilityManager;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
 import org.labkey.api.specimen.model.SpecimenComment;
-import org.labkey.specimen.query.SpecimenQueryView;
 import org.labkey.api.specimen.security.permissions.EditSpecimenDataPermission;
 import org.labkey.api.specimen.security.permissions.ManageRequestSettingsPermission;
 import org.labkey.api.specimen.security.permissions.RequestSpecimensPermission;
@@ -152,7 +148,10 @@ import org.labkey.specimen.AmbiguousLocationException;
 import org.labkey.specimen.RequestEventType;
 import org.labkey.specimen.RequestedSpecimens;
 import org.labkey.specimen.SpecimenManager;
+import org.labkey.specimen.SpecimenRequestException;
 import org.labkey.specimen.SpecimenRequestManager;
+import org.labkey.specimen.SpecimenRequestStatus;
+import org.labkey.specimen.importer.RequestabilityManager;
 import org.labkey.specimen.model.ExtendedSpecimenRequestView;
 import org.labkey.specimen.model.SpecimenRequestActor;
 import org.labkey.specimen.model.SpecimenRequestEvent;
@@ -162,6 +161,7 @@ import org.labkey.specimen.notifications.NotificationRecipientSet;
 import org.labkey.specimen.pipeline.SpecimenArchive;
 import org.labkey.specimen.pipeline.SpecimenBatch;
 import org.labkey.specimen.query.SpecimenEventQueryView;
+import org.labkey.specimen.query.SpecimenQueryView;
 import org.labkey.specimen.query.SpecimenRequestQueryView;
 import org.labkey.specimen.requirements.SpecimenRequest;
 import org.labkey.specimen.requirements.SpecimenRequestRequirement;
