+++ specimen/src/org/labkey/specimen/actions/ViewRequestsHeaderBean.java
@@ -1,8 +1,8 @@
 package org.labkey.specimen.actions;
 
-import org.labkey.api.specimen.SpecimenRequestStatus;
 import org.labkey.api.view.ViewContext;
 import org.labkey.specimen.SpecimenRequestManager;
+import org.labkey.specimen.SpecimenRequestStatus;
 import org.labkey.specimen.query.SpecimenRequestQueryView;
 
 import java.util.List;
