+++ specimen/src/org/labkey/specimen/importer/SpecimenImporter.java
@@ -27,7 +27,29 @@ import org.junit.Before;
 import org.junit.Test;
 import org.labkey.api.collections.CaseInsensitiveHashMap;
 import org.labkey.api.collections.CaseInsensitiveHashSet;
-import org.labkey.api.data.*;
+import org.labkey.api.data.BaseColumnInfo;
+import org.labkey.api.data.ColumnInfo;
+import org.labkey.api.data.Container;
+import org.labkey.api.data.DbSchema;
+import org.labkey.api.data.DbScope;
+import org.labkey.api.data.DbSequence;
+import org.labkey.api.data.DbSequenceManager;
+import org.labkey.api.data.JdbcType;
+import org.labkey.api.data.Parameter;
+import org.labkey.api.data.Results;
+import org.labkey.api.data.ResultsImpl;
+import org.labkey.api.data.RuntimeSQLException;
+import org.labkey.api.data.SQLFragment;
+import org.labkey.api.data.SimpleFilter;
+import org.labkey.api.data.Sort;
+import org.labkey.api.data.SqlExecutor;
+import org.labkey.api.data.SqlSelector;
+import org.labkey.api.data.Table;
+import org.labkey.api.data.TableInfo;
+import org.labkey.api.data.TableResultSet;
+import org.labkey.api.data.TableSelector;
+import org.labkey.api.data.TempTableInfo;
+import org.labkey.api.data.UpdateableTableInfo;
 import org.labkey.api.data.dialect.SqlDialect;
 import org.labkey.api.dataiterator.DataIterator;
 import org.labkey.api.dataiterator.DataIteratorBuilder;
@@ -56,17 +78,13 @@ import org.labkey.api.reader.ColumnDescriptor;
 import org.labkey.api.reader.DataLoader;
 import org.labkey.api.reader.Readers;
 import org.labkey.api.security.User;
-import org.labkey.specimen.SpecimenColumns;
 import org.labkey.api.specimen.SpecimenEvent;
 import org.labkey.api.specimen.SpecimenEventDateComparator;
 import org.labkey.api.specimen.SpecimenEventManager;
 import org.labkey.api.specimen.SpecimenMigrationService;
 import org.labkey.api.specimen.SpecimenSchema;
-import org.labkey.specimen.SpecimenRequestManager;
-import org.labkey.specimen.SpecimenTableManager;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.importer.EventVialRollup;
-import org.labkey.api.specimen.importer.RequestabilityManager;
 import org.labkey.api.specimen.importer.RollupInstance;
 import org.labkey.api.specimen.location.LocationCache;
 import org.labkey.api.specimen.location.LocationManager;
@@ -98,6 +116,10 @@ import org.labkey.api.util.ResultSetUtil;
 import org.labkey.api.util.StringUtilsLabKey;
 import org.labkey.api.util.TestContext;
 import org.labkey.api.writer.VirtualFile;
+import org.labkey.specimen.SpecimenColumns;
+import org.labkey.specimen.SpecimenManager;
+import org.labkey.specimen.SpecimenRequestManager;
+import org.labkey.specimen.SpecimenTableManager;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -893,7 +915,7 @@ public class SpecimenImporter extends SpecimenTableManager
                     _iTimer.setPhase(ImportPhases.GetProcessingLocationId);
                     Integer processingLocation = LocationManager.get().getProcessingLocationId(dateOrderedEvents);
                     _iTimer.setPhase(ImportPhases.GetFirstProcessedBy);
-                    String firstProcessedByInitials = SpecimenEventManager.get().getFirstProcessedByInitials(dateOrderedEvents);
+                    String firstProcessedByInitials = SpecimenManager.get().getFirstProcessedByInitials(dateOrderedEvents);
                     _iTimer.setPhase(ImportPhases.GetCurrentLocationId);
                     Integer currentLocation = LocationManager.get().getCurrentLocationId(dateOrderedEvents);
