+++ specimen/src/org/labkey/specimen/importer/SpecimenSettingsImporter.java
@@ -20,8 +20,6 @@ import org.labkey.api.attachments.AttachmentService;
 import org.labkey.api.data.Container;
 import org.labkey.api.security.GroupManager;
 import org.labkey.api.security.SecurityManager;
-import org.labkey.api.specimen.SpecimenRequestStatus;
-import org.labkey.api.specimen.importer.RequestabilityManager;
 import org.labkey.api.specimen.location.LocationCache;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.settings.DisplaySettings;
@@ -35,6 +33,7 @@ import org.labkey.api.writer.VirtualFile;
 import org.labkey.security.xml.GroupType;
 import org.labkey.specimen.SpecimenRequestManager;
 import org.labkey.specimen.SpecimenRequestManager.SpecimenRequestInput;
+import org.labkey.specimen.SpecimenRequestStatus;
 import org.labkey.specimen.actions.ManageReqsBean;
 import org.labkey.specimen.model.SpecimenRequestActor;
 import org.labkey.specimen.requirements.RequirementType;
