+++ specimen/src/org/labkey/specimen/view/NotificationBean.java
@@ -4,7 +4,6 @@ import org.jetbrains.annotations.NotNull;
 import org.labkey.api.attachments.Attachment;
 import org.labkey.api.data.Container;
 import org.labkey.api.security.User;
-import org.labkey.api.specimen.SpecimenRequestStatus;
 import org.labkey.api.specimen.location.LocationManager;
 import org.labkey.api.study.Location;
 import org.labkey.api.study.SpecimenUrls;
@@ -12,6 +11,7 @@ import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.ViewContext;
 import org.labkey.specimen.SpecimenRequestManager;
+import org.labkey.specimen.SpecimenRequestStatus;
 import org.labkey.specimen.model.SpecimenRequestEvent;
 import org.labkey.specimen.notifications.DefaultRequestNotification;
