+++ specimen/src/org/labkey/specimen/view/configRequestabilityRules.jsp
@@ -16,10 +16,10 @@
  */
 %>
 <%@ page import="org.labkey.api.data.Container"%>
-<%@ page import="org.labkey.api.specimen.importer.RequestabilityManager"%>
 <%@ page import="org.labkey.api.study.StudyUrls"%>
-<%@ page import="org.labkey.api.view.ActionURL" %>
+<%@ page import="org.labkey.api.view.ActionURL"%>
 <%@ page import="org.labkey.api.view.template.ClientDependencies" %>
+<%@ page import="org.labkey.specimen.importer.RequestabilityManager" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%!
