+++ specimen/src/org/labkey/specimen/view/manageStatuses.jsp
@@ -15,15 +15,15 @@
  * limitations under the License.
  */
 %>
-<%@ page import="org.labkey.api.specimen.SpecimenRequestStatus"%>
 <%@ page import="org.labkey.api.specimen.settings.SettingsManager"%>
 <%@ page import="org.labkey.api.specimen.settings.StatusSettings"%>
 <%@ page import="org.labkey.api.study.Study"%>
 <%@ page import="org.labkey.api.study.StudyUrls"%>
-<%@ page import="org.labkey.api.view.ActionURL" %>
+<%@ page import="org.labkey.api.view.ActionURL"%>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.specimen.SpecimenRequestManager" %>
+<%@ page import="org.labkey.specimen.SpecimenRequestStatus" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController.DeleteStatusAction" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController.ManageStatusOrderAction" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController.ManageStatusesAction" %>
